use error message always
It's hard to figure out what cause the revert if there
is no message. This would be an issue for whoever integrates
clients with the smart contract system, but also for us
when attempting to write tests.

I chose to write tests with the "<contract name>: <error msg>"
format. Some contracts already do that, others don't. We can
transition into using this format consistently across the whole
code base over time.